board/espressif/esp32s3: add persistent /etc
Add /etc/fstab to the board file system overlay with /etc in it so that
persistent configuration could be maintained on the device.
Add a script that makes JFFS2 file system image with /etc contents.
Enable building JFFS2 image. Although the result is called rootfs.jffs2,
in fact it's only /etc.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>